Designate lab 1 a pre-lab
We're seeing a lot of folks come to the loft to this workshop and
getting stuck on setup steps like installing with AWS CLI. By
designating lab 1 a pre-lab, we can promote this workshop with the
requirement that the pre-lab be completed prior to the workshop to
give students a head start on getting their environment configured.

This has the additional benefit of moving the CloudFront distribution
creation step to be done prior to the workshop starting as that can
sometimes take upwards of 30 minutes to provision.

- Decrement subsequent lab numbers
- Rename images
- Update copyright notice